refactor: move ClusterCommand to a separate module
