+++ src/services/cluster/actor.rs
@@ -3,13 +3,7 @@ use std::collections::BTreeMap;
 use tokio::{net::TcpStream, sync::mpsc::Receiver};
 
 use crate::make_smart_pointer;
-
-pub enum ClusterCommand {
-    AddPeer(PeerAddr, TcpStream),
-    RemovePeer(PeerAddr),
-    GetPeer(PeerAddr),
-    GetPeers(tokio::sync::oneshot::Sender<Vec<PeerAddr>>),
-}
+use crate::services::cluster::command::ClusterCommand;
 
 pub struct ClusterActor {
     pub peers: BTreeMap<PeerAddr, Connected>,
